# storage/api-docs/api-docs.json
{
    "swagger": "2.0",
    "info": {
        "title": "User - Lendtick",
        "contact": {
            "email": "faujiakbar@gmail.com"
        },
        "version": "1.0.0"
    },
    "basePath": "/",
    "schemes": [
        "http",
        "https"
    ],
    "paths": {
...
        "/profile/delete-address-by-user": {
            "post": {
                "tags": [
                    "Profile"
                ],
                "summary": "Notifications",
                "description": "Send Notification Lendtick",
                "operationId": "Address",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "in": "body",
                        "name": "body",
                        "description": "",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "properties": { 
                                "id_user_address": {
                                    "type": "integer",
                                    "example": 8
                                }
                            },
                            "required": ["id_user_address"],
                            "$schema": "http://json-schema.org/draft-04/schema#"
                        }
                    }],
                "responses": {
                    "200": {
                        "description": ""
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            }
        }, 
        "/profile/get-address-by-user": {
            "get": {
                "tags": [
                    "Profile"
                ],
                "summary": "Get List Address of User",
                "description": "Get User Address of User",
                "operationId": "Address",
                "produces": [
                    "application/json"
                ],
                "responses": {
                    "200": {
                        "description": "successful"
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            }
        },
...
    },
...
    "securityDefinitions": {
        "Bearer": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}